Split superwidget into separate files
