+++ demo_widget.html
@@ -13,7 +13,6 @@
 </span>
 
 <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
-<script src="dist/dev/yesgraph.js"></script>
 <script src="dist/dev/yesgraph-invites.js"></script>
 <script>
 $(document).on("installed.yesgraph.sdk", function(evt) {
