+++ gulp/tasks/minify.js
@@ -7,13 +7,40 @@ var sourcemaps = require("gulp-sourcemaps");
 var debug = require('gulp-debug');
 var config = require("../config");
 
-gulp.task("minify:js", function() {
-    return gulp.src(config.tasks.minifyScripts.files, {base: config.src.root})
+var browserify = require('browserify');
+var buffer = require('vinyl-buffer');
+var gutil = require('gulp-util');
+var source = require('vinyl-source-stream');
+
+var watchify = require('watchify');
+var babel = require('babelify');
+var through = require("through2");
+
+gulp.task("minify", ["minify:js", "minify:css"]);
+
+gulp.task("minify:js", ["minify:js:sdk", "minify:js:superwidget"]);
+
+gulp.task("minify:js:superwidget", function(){
+    var bundler = browserify({
+        entries: "./src/dev/superwidget.js",
+        debug: true
+    }).transform('babelify', { presets: ['es2015' ]});
+    return bundler.bundle()
+        .pipe(source('yesgraph-invites.js'))
+        .pipe(buffer())
         .pipe(debug({title: "minify:js"}))
-        .pipe(sourcemaps.init())
-        .pipe(uglify({preserveComments: "license"}))
-        .pipe(rename({suffix: ".min"}))
-        .pipe(sourcemaps.write("."))
+        .pipe(gulp.dest(config.dest.root));
+});
+
+gulp.task("minify:js:sdk", function(){
+    var bundler = browserify({
+        entries: "./src/dev/sdk.js",
+        debug: true
+    }).transform('babelify', { presets: ['es2015' ]});
+    return bundler.bundle()
+        .pipe(source('yesgraph.js'))
+        .pipe(buffer())
+        .pipe(debug({title: "minify:js"}))
         .pipe(gulp.dest(config.dest.root));
 });
 
@@ -25,4 +52,13 @@ gulp.task("minify:css", function() {
         .pipe(gulp.dest(config.dest.root));
 });
 
-gulp.task("minify", ["minify:js", "minify:css"]);
+
+// SOURCEMAPS SNIPPET
+// 
+// .pipe(sourcemaps.init())
+// .pipe(babel({
+//     presets: ['es2015']
+// }))
+// .pipe(uglify({preserveComments: "license"}))
+// .pipe(rename({suffix: ".min"}))
+// .pipe(sourcemaps.write("."))
