+++ gulp/tasks/watch.js
@@ -1,5 +1,6 @@
 "use strict";
 var gulp = require("gulp");
+var gutil = require('gulp-util');
 var config = require("../config");
 
 gulp.task("_js", ["version"], function(){
